refactor: 수익계산기의 위치를 service 계층으로 변경
